+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Added
+
+- `Subscribe()` now supports subscribing to interface types. The function will
+  receive any messages that implement the interface.
+
 ## [0.2.0] - 2024-07-31
 
 This release abandons the "component" terminology and simply refers to the